\name{DBfit-package}
\alias{DBfit-package}
\alias{DBfit}
\docType{package}
\title{
A Double Bootstrap Method for Analyzing Linear Models With Autoregressive Errors
}
\description{
\packageDescription{DBfit}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{DBfit}
\packageIndices{DBfit}
}
\author{
\packageAuthor{DBfit}

Maintainer: \packageMaintainer{DBfit}
}
\references{
McKnight, S. D., McKean, J. W., and Huitema, B. E. (2000). A double bootstrap method
to analyze linear models with autoregressive error terms. Psychological methods,
5 (1), 87.
Shaofeng Zhang (2017). Ph.D. Dissertation.
}


