\name{print.dbfit}
\alias{print.dbfit}
\alias{print.summary.dbfit}

\title{
DBfit Internal Print Functions
}
\description{
These functions print the output in a user-friendly manner using the internal R function \code{print}.
}
\usage{
\method{print}{dbfit}(x, ...)
\method{print}{summary.dbfit}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An object to be printed
}
  \item{\dots}{
additional arguments to be passed to \code{print}
}
}
\seealso{
\code{\link{dbfit}}, \code{\link{summary.dbfit}}
}