\name{testdata}
\alias{testdata}
\docType{data}
\title{
testdata
}
\description{
This data serves as a test data.
}
\usage{data("testdata")}
\format{
  A data frame with 40 observations. First 4 columns make up the design matrix, while the last column is the response variable. This data is generated by a two-phase design, with autoregressive order being one, autoregressive coefficient being 0.6 and all regression coefficients being 0. Both the first and second phase have 20 observations.
}
\examples{
data(testdata)
}
\keyword{datasets}
